% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_gru_thesis.R
\name{pdf_gru_thesis}
\alias{pdf_gru_thesis}
\title{GRU thesis template for pdf output}
\usage{
pdf_gru_thesis(..., keep_tex = TRUE, toc = FALSE,
  number_sections = FALSE)
}
\arguments{
\item{...}{Arguments to \code{bookdown::pdf_document2}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{number_sections}{Whether to number section headers: if \code{TRUE},
figure/table numbers will be of the form \code{X.i}, where \code{X} is the
current first-level section number, and \code{i} is an incremental number
(the i-th figure/table); if \code{FALSE}, figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference
section headers in this case.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
GRU thesis template for pdf output
}
\examples{

\dontrun{
library(gruRmarkdown)

# Create draft
rmarkdown::draft(file = "my_thesis.Rmd", create_dir = TRUE,
  template = "gru_thesis", package = "gruRmarkdown",  edit = FALSE)

# Render thesis PDF
rmarkdown::render(input = "my_thesis/my_thesis.Rmd", 
  output_format = c("gruRmarkdown::pdf_gru_thesis"))

}

}
