% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fineprint_brief.R
\name{fineprint_brief}
\alias{fineprint_brief}
\title{FINEPRINT Brief template}
\usage{
fineprint_brief(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
FINEPRINT Brief template
}
\examples{

\dontrun{
library(iiasaRmarkdown)

rmarkdown::draft(file = "my_brief.Rmd", create_dir = TRUE,
  template = "fineprint_brief", package = "gruRmarkdown")

rmarkdown::render(input = "my_brief.Rmd")

}

}
